// src/types.ts
export interface RoleProps {
  roleName: string;
  description: string;
  companyId: number;
  id: number;
  isActive: number;
}